feat: add highlight if several buttons are pressed, split the code into modules
