+++ index.html
@@ -12,6 +12,8 @@
       src="https://kit.fontawesome.com/5d2ab4825d.js"
       crossorigin="anonymous"
     ></script>
-    <script src="index.js"></script>
+    <script src="index.js" type="module"></script>
+    <script src="keys.js" type="module"></script>
+    <script src="keyboardEventHandler.js" type="module"></script>
   </body>
 </html>
